requirements.txt
streamlit
matplotlib
numpy
skimage
PIL
scipy.ndimage
skimage.measure
pandas